Properly functioning Nav menus!!
The trick here was concealing (on md and lg) some buttons near the
existing dropdown list.  The mobile menu will render those as a
dropdown, and we suppress the lg dropdown, b/c it was off-canvas and
required an extra click to discover.  This version should be more
mobile-friendly

The JS work is to correct a known issue with the Topbar module, and
you should now be able to click/tap outside the menu to get it to
collapse.  This was not happening before.

Much styling work to do, but functionality appears good.