Adding News ticker, with 5s rotation
This is using JQuery and I have added some custom functions to track
the size of the viewport, because there is no need to have this
Javascript running on devices where the entire news feed is not
displayed.  Note the use of console logging statemtents, to be
deleted at a later date.