Z-index changes across the board
We're using variables for things like the hero div and the
sharing sidebar, and the directives should be in locations
useful for those working on other tasks.  Previous issues
regarding the hero image and overlays darkening headings have
been dealt with in this patch.

Also of note: removal of unnecessary 'spacer' divs in grid.
These foundation classes should bring about the same behavior.