Fixes #10
This Javascript applies to all project pages.
Testing: please find anchor links (links within the page, and links
to locations in other pages on our site) and see what results.

Ideally, any link with a '#' in it should be affected by this
Javascript function, and will set the targeted element 5px below
the navigation bar.