Another attempt at fixing #20
iPhone 6 on iOS 9 appears to be using a viewport approx. 5% larger
than specs.  The media query here correctly activates on the 6, both
real and simulated (sim is using iOS 8.3, but does now exhibit the
undesirable whitespace at R)

This styling is a brute force attempt to get the page to occupy
100% of the viewport width.  If no other solution is found before
launch, we can duplicate this for the 5, 6s, etc as needed.