Adding Galano Grotesque fonts
We purchased a license for three weight, regular and italic
and I have corrected the metadata in the TrueType versions,
so there are TrueType XML (.ttx) files here.  They are not
needed for web deployment.

This branch will be used as a holding place, in case we run
into trouble serving these files on the web, or using them
in graphic design.