Popup email solicitation with grace period
This is an implementation of Foundations Modal reveal
functionality, with callbacks to ensure there is no
conflict with the video modal (on use on issue pages and
possibly more to come)

Modals (including video) are now styled as position:fixed
to ensure they remain on the viewport despite user scrolling

LOTS of new Javascript here, including JS-cookie, a JQuery plugin
for easy manipulation of cookies.

Defaults:

1.  Setting a grace period (1 day for now) to prevent modals from
revealing on scroll.

2.  Use of LocalStorage is preferred over cookies, with cookies as
a fallback if local device unable to handle the former.

3.  Console logging of program behavior during testing.

There is an additional Jade template, in case we design a similar
modal popup to solicit donations.  Howard has agreed to go with
email solicitation for now.  Styling in scss covers both identified
elements, globally for the entire site.

Timing of visitor relief from popups is global for the FY site.