news ticker improvements: ensuring display values
These stylesheet and js changes are designed to ensure that the
proper list items are displayed when animations occur, and concealed
when they are not wanted.  We have been seeing situations where
more than one item is displaying at once, and this setup is aimed
at ensuring the first list item is displayed on page load, and
siblings are out of the way.